requirements.txt

python-environ
langchain
langchain_experimental
sentence_transformers
langchain_openai
langchain_ollama
pypdf
unstructured
python-docx
csvreader
faiss-cpu


